add warning if ca backup doesn't have necessary files to import
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@8130 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart/src/org/wcs/smart/ca/in/CaImporter.java
@@ -111,6 +111,9 @@ public class CaImporter {
 			try(ZipFile archiveFile = new ZipFile(file.normalize().toAbsolutePath().toFile())){
 				ZipArchiveEntry zipEntry = archiveFile.getEntry(ICaDataExportEngine.CA_INFO_FILENAME);
 				Path p = temp.resolve(ICaDataExportEngine.CA_INFO_FILENAME);
+				if (!Files.exists(p)) {
+					throw new Exception(Messages.CaImporter_NotCaFile);
+				}
 				Files.createDirectories(p.getParent());
 				Files.copy(archiveFile.getInputStream(zipEntry), p);
